refactor(prisma): extract the BigInt money map to lib/prisma-extensions.ts
lib/prisma.ts went 294 -> 116 lines: the ~190-line #780/#781 result-extension
map (the real bloat) now lives in lib/prisma-extensions.ts, leaving prisma.ts as
just the connection/timeout/client setup. Zero behaviour change — Db/Tx types are
still derived from the composed client (tsc --noEmit clean, no excessive-stack-depth).
Point the #780 drift test at the new file.
